feat(skills): modalize detail and add cli install panel
